+++ PowerSync/PowerSync.Common/CHANGELOG.md
@@ -1,6 +1,7 @@
 # PowerSync.Common Changelog
 
 ## 0.0.6-alpha.1
+- Dropping support for the legacy C# sync implementation.
 - Add `trackPreviousValues` option on `TableOptions` which sets `CrudEntry.PreviousValues` to previous values on updates.
 - Add `trackMetadata` option on `TableOptions` which adds a `_metadata` column that can be used for updates. The configured metadata is available through `CrudEntry.Metadata`.
 - Add `ignoreEmptyUpdates` option on `TableOptions` which skips creating CRUD entries for updates that don't change any values.