+++ PowerSync/PowerSync.Common/CHANGELOG.md
@@ -1,8 +1,14 @@
 # PowerSync.Common Changelog
 
 ## 0.0.6-alpha.1
-- Added ability to specify `AppMetadata` for sync/stream requests.
-
+- Dropping support for the legacy C# sync implementation.
+- Add `trackPreviousValues` option on `TableOptions` which sets `CrudEntry.PreviousValues` to previous values on updates.
+- Add `trackMetadata` option on `TableOptions` which adds a `_metadata` column that can be used for updates. The configured metadata is available through `CrudEntry.Metadata`.
+- Add `ignoreEmptyUpdates` option on `TableOptions` which skips creating CRUD entries for updates that don't change any values.
+- Reporting progress information about downloaded rows. Sync progress is available through `SyncStatus.DownloadProgress()`.
+- Support bucket priorities.
+- Report `PriorityStatusEntries` on `SyncStatus`.
+- Added ability to specify `AppMetadata` for sync/stream requests
 Note: This requires a PowerSync service version `>=1.17.0` in order for logs to display metadata.
 
 ```csharp