+++ PowerSync/PowerSync.Common/CHANGELOG.md
@@ -1,6 +1,7 @@
 # PowerSync.Common Changelog
 
 ## 0.0.6-alpha.1
+- Updated to the lastest version (0.4.10) of the core extension.
 - Dropping support for the legacy C# sync implementation.
 - Add `trackPreviousValues` option on `TableOptions` which sets `CrudEntry.PreviousValues` to previous values on updates.
 - Add `trackMetadata` option on `TableOptions` which adds a `_metadata` column that can be used for updates. The configured metadata is available through `CrudEntry.Metadata`.
@@ -10,7 +11,7 @@
 - Report `PriorityStatusEntries` on `SyncStatus`.
 
 ## 0.0.5-alpha.1
-- Using the latest (0.4.9) version of the core extension, it introduces support for the Rust Sync implementation and also makes it the default - users can still opt out and use the legacy C# sync implementation as option when calling `connect()`.
+- Using the latest version (0.4.9) of the core extension, it introduces support for the Rust Sync implementation and also makes it the default - users can still opt out and use the legacy C# sync implementation as option when calling `connect()`.
 
 ## 0.0.4-alpha.1
 - Fixed MAUI issues related to extension loading when installing package outside of the monorepo.